Merges the experimental_table_slice back into arrow_table_slice
Builders for both `arrow::v1` and `arrow::experimental` are still in
place to faciliate unit testing for both implemenations.
`arrow::v1` is still the default implementation, so this change doesn't
touch any of the default behavior.
