--- libvast/vast/experimental_table_slice.hpp
@@ -1,120 +1,0 @@
-//    _   _____   __________
-//   | | / / _ | / __/_  __/     Visibility
-//   | |/ / __ |_\ \  / /          Across
-//   |___/_/ |_/___/ /_/       Space and Time
-//
-// SPDX-FileCopyrightText: (c) 2021 The VAST Contributors
-// SPDX-License-Identifier: BSD-3-Clause
-
-#pragma once
-
-#include "vast/fwd.hpp"
-
-#include "vast/table_slice.hpp"
-
-#include <arrow/type_fwd.h>
-#include <caf/meta/type_name.hpp>
-
-#include <memory>
-
-namespace vast {
-
-/// Additional state needed for the implementation of Arrow-encoded table slices
-/// that cannot easily be accessed from the underlying FlatBuffers table
-/// directly.
-struct experimental_table_slice_state {
-  /// The deserialized table layout.
-  type layout;
-
-  /// The deserialized Arrow Record Batch.
-  std::shared_ptr<arrow::RecordBatch> record_batch;
-
-  /// Mapping from column offset to nested Arrow array
-  arrow::ArrayVector array_index;
-};
-
-/// A table slice that stores elements encoded in the [Arrow](https://arrow.org)
-/// format. The implementation stores data in column-major order.
-class experimental_table_slice final {
-public:
-  // -- constructors, destructors, and assignment operators --------------------
-
-  /// Constructs an Arrow-encoded table slice from a FlatBuffers table.
-  /// @param slice The encoding-specific FlatBuffers table.
-  /// @param parent The surrounding chunk.
-  experimental_table_slice(const fbs::table_slice::arrow::experimental& slice,
-                           const chunk_ptr& parent) noexcept;
-
-  /// Destroys a Arrow-encoded table slice.
-  ~experimental_table_slice() noexcept;
-
-  // -- properties -------------------------------------------------------------
-
-  /// Whether the most recent version of the encoding is used.
-  inline static constexpr bool is_latest_version = true;
-
-  /// The encoding of the slice.
-  inline static constexpr enum table_slice_encoding encoding
-    = table_slice_encoding::experimental;
-
-  /// @returns The table layout.
-  [[nodiscard]] const type& layout() const noexcept;
-
-  /// @returns The number of rows in the slice.
-  [[nodiscard]] table_slice::size_type rows() const noexcept;
-
-  /// @returns The number of columns in the slice.
-  [[nodiscard]] table_slice::size_type columns() const noexcept;
-
-  // -- data access ------------------------------------------------------------
-
-  /// Appends all values in column `column` to `index`.
-  /// @param `offset` The offset of the table slice in its ID space.
-  /// @param `column` The index of the column to append.
-  /// @param `index` the value index to append to.
-  void append_column_to_index(id offset, table_slice::size_type column,
-                              value_index& index) const;
-
-  /// Retrieves data by specifying 2D-coordinates via row and column.
-  /// @param row The row offset.
-  /// @param column The column offset.
-  /// @pre `row < rows() && column < columns()`
-  [[nodiscard]] data_view
-  at(table_slice::size_type row, table_slice::size_type column) const;
-
-  /// Retrieves data by specifying 2D-coordinates via row and column.
-  /// @param row The row offset.
-  /// @param column The column offset.
-  /// @param t The type of the value to be retrieved.
-  /// @pre `row < rows() && column < columns()`
-  [[nodiscard]] data_view
-  at(table_slice::size_type row, table_slice::size_type column,
-     const type& t) const;
-
-  /// @returns The import timestamp.
-  [[nodiscard]] time import_time() const noexcept;
-
-  /// Sets the import timestamp.
-  void import_time(time import_time) noexcept;
-
-  /// @returns A shared pointer to the underlying Arrow Record Batch.
-  [[nodiscard]] std::shared_ptr<arrow::RecordBatch>
-  record_batch() const noexcept;
-
-  /// Provides a data column for an offset.
-  /// @param column The column index
-  /// @returns The arrow array backing the (flattened) column
-  [[nodiscard]] std::shared_ptr<arrow::Array>
-  column_array(table_slice::size_type column) const noexcept;
-
-private:
-  // -- implementation details -------------------------------------------------
-
-  /// A const-reference to the underlying FlatBuffers table.
-  const fbs::table_slice::arrow::experimental& slice_;
-
-  /// Additional state needed for the implementation.
-  experimental_table_slice_state state_;
-};
-
-} // namespace vast
